Buffer is now a header only library. Tiddied header imports in source files.
