+++ lib/hydra_server.cc
@@ -1,7 +1,5 @@
 #include "hydra/hydra_server.h"
-#include "hydra/util/udp.h"
 
-#include <boost/algorithm/string.hpp>
 
 namespace hydra {
